+++ docs/CHANGE_LOG.md
@@ -1,7 +1,9 @@
 # Heat.js - Change Log:
 
 ## Version 0.2.0:
-- 
+
+#### **Binding Options - Custom Triggers:**
+- Added a new binding option called "onRefresh", which states an event that should be triggered when a heatmap is refreshed.
 
 <br>