+++ docs/CHANGE_LOG.md
@@ -2,6 +2,9 @@
 
 ## Version 0.2.0:
 
+#### **Binding Options:**
+- Added a new binding option called "dayToolTipText", which states the tooltip text that should be shown for a day.
+
 #### **Binding Options - Custom Triggers:**
 - Added a new binding option called "onRefresh", which states an event that should be triggered when a heatmap is refreshed.
 - Added a new binding option called "onBeforeRender", which states an event that should be triggered before an element is rendered.