+++ docs/CHANGE_LOG.md
@@ -12,10 +12,11 @@
 
 #### **Public Functions:**
 - Added new public function "removeDate()", which will remove a specific date from a heat map (or decrease the date if more than one exists).
-- Added new public function "reset()", which will remove all the dates from a the heat map.
+- Added new public function "reset()", which will remove all the dates from a heat map.
 
 #### **Documentation:**
 - Improved documentation in the main README files.
+- Fixed some spelling/grammar mistakes.
 
 <br>